refactor(mobility): use Base64ImageSchema
